+++ contracts/libraries/PremiumPosition.sol
@@ -11,7 +11,7 @@ library PremiumPosition {
     uint256 beginPremiumRate;
     //Thao@TODO: pack ownerIndex and lastTick in uint256 for saving gas
     uint256 ownerIndex;
-    uint24 lastTick;
+    uint32 lastTick;
   }
 
   function get(
